Created FrameBasedSurface from PSurfaceAWT.
